921 - DEVEX - refactored the use of createSilentApplicationContext to make it reusable
